refactor: move icon components to a dedicated icons directory
