+++ src/routes/+page.svelte
@@ -2,7 +2,7 @@
 	import IconCustomizer from '../components/icons/IconCustomizer.svelte';
 	import '../app.css';
 	import Sidenav from '../components/Sidenav.svelte';
-	import IconPreview from '../components/IconPreview.svelte';
+	import IconPreview from '../components/icons/IconPreview.svelte';
 	import Toolbar from '../components/Toolbar/Toolbar.svelte';
 	import ShortcutsBar from '../components/ShortcutsBar.svelte';
